garden_advisor_project/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic
svgwrite
fastapi
uvicorn[standard] # [standard] inkluderar gunicorn/uvloop för produktion
python-dotenv
supabase
google-cloud-aiplatform # Eller google-generativeai
svgwrite
Pillow # Om du använder den för bildvalidering eller bearbetning
# Eventuellt andra paket du lagt till